Mudanças Importantes
Adiciona classe que servirá de base para renderizar imagens. Adicionado componentes para ECS: RigidBody, Transform. Adiciona começo do componente Collider. Remove GameInstance para reformulação

+++ glw/fps-counter.cpp
@@ -1,6 +1,6 @@
 #include <glw/utils/game/fps-counter.hpp>
 
-void glw::utils::game::_FPSCounter::_Load(double dt) noexcept {
+void glw::utils::game::_FPSCounter::_Load(float dt) noexcept {
     this->FPSCountTimeout -= dt;
 
     if (this->FPSCountTimeout <= 0) {
